Refactored participant-status heavily
