+++ server/services/participant-status/get-statuses.ts
@@ -1,0 +1,13 @@
+import { dbClient, collections } from '../../db';
+
+export const getParticipantHiredStatuses = async (participantId: number) => {
+  const statuses = await dbClient.db[collections.PARTICIPANTS_STATUS].find({
+    participant_id: participantId,
+    status: 'hired',
+    current: true,
+  });
+  if (statuses.length === 0) {
+    throw new Error('Participant is not hired');
+  }
+  return statuses;
+};
